extract feedback and notes section
