+++ src/app/[orgSlug]/study/[studyId]/review/post-feedback-view.tsx
@@ -2,13 +2,11 @@
 
 import { PageBreadcrumbs } from '@/components/page-breadcrumbs'
 import type { ReviewDecision } from '@/database/types'
+import { FeedbackAndNotesSection } from '@/components/study/feedback-and-notes'
 import { Routes } from '@/lib/routes'
-import { extractTextFromLexical } from '@/lib/word-count'
-import { Anchor, Box, Button, Collapse, Divider, Group, Paper, Stack, Text, Title } from '@mantine/core'
-import { CaretDownIcon, CaretUpIcon } from '@phosphor-icons/react'
+import { Box, Button, Divider, Group, Stack, Text, Title } from '@mantine/core'
 import dayjs from 'dayjs'
 import { useRouter } from 'next/navigation'
-import { useState } from 'react'
 import type { ProposalFeedbackEntry, SelectedStudy } from '@/server/actions/study.actions'
 import { ProposalSection } from './proposal-section'
 
@@ -57,16 +55,6 @@ function formatDate(date: Date | string): string {
     return dayjs(date).format('MMM DD, YYYY')
 }
 
-function entryTitle(entry: ProposalFeedbackEntry): string {
-    return entry.entryType === 'REVIEWER-FEEDBACK' ? 'Reviewer feedback' : 'Resubmission note'
-}
-
-// TODO(OTTER-491): swap plain-text rendering for a Lexical viewer once the editor
-// component lands; share node config with the textarea.
-function entryBodyText(entry: ProposalFeedbackEntry): string {
-    return extractTextFromLexical(JSON.stringify(entry.body))
-}
-
 function DecisionBanner({ decision }: { decision: ReviewDecision }) {
     const { banner } = DECISION_COPY[decision]
     return (
@@ -80,101 +68,6 @@ function DecisionBanner({ decision }: { decision: ReviewDecision }) {
 
 function FullProposalDropdown({ orgSlug, study }: { orgSlug: string; study: SelectedStudy }) {
     return <ProposalSection orgSlug={orgSlug} study={study} initialExpanded={false} />
-}
-
-function ToggleCaret({ isExpanded }: { isExpanded: boolean }) {
-    const Icon = isExpanded ? CaretUpIcon : CaretDownIcon
-    return <Icon size={14} />
-}
-
-type FeedbackEntryProps = {
-    entry: ProposalFeedbackEntry
-    isExpanded: boolean
-    onToggle: () => void
-}
-
-function FeedbackEntry({ entry, isExpanded, onToggle }: FeedbackEntryProps) {
-    const title = entryTitle(entry)
-    const body = entryBodyText(entry)
-    const date = formatDate(entry.createdAt)
-
-    return (
-        <Stack gap="sm" data-testid={`feedback-entry-${entry.id}`}>
-            <Group justify="space-between" align="center">
-                <Stack gap={2}>
-                    <Text fw={600}>{title}</Text>
-                    <Text size="sm" c="gray.7">
-                        {entry.authorName} · {date}
-                    </Text>
-                </Stack>
-                <Anchor
-                    component="button"
-                    onClick={onToggle}
-                    c="blue"
-                    size="sm"
-                    data-testid={`feedback-toggle-${entry.id}`}
-                    aria-expanded={isExpanded}
-                >
-                    <ToggleCaret isExpanded={isExpanded} />
-                </Anchor>
-            </Group>
-            <Collapse in={isExpanded}>
-                <Text size="sm" data-testid={`feedback-body-${entry.id}`} style={{ whiteSpace: 'pre-wrap' }}>
-                    {body}
-                </Text>
-            </Collapse>
-        </Stack>
-    )
-}
-
-function useExpandedEntries(entries: ProposalFeedbackEntry[]) {
-    const [expandedIds, setExpandedIds] = useState<Set<string>>(() => {
-        const latest = entries[0]?.id
-        return latest ? new Set([latest]) : new Set()
-    })
-
-    const toggle = (id: string) => {
-        setExpandedIds((prev) => {
-            const next = new Set(prev)
-            if (next.has(id)) {
-                next.delete(id)
-            } else {
-                next.add(id)
-            }
-            return next
-        })
-    }
-
-    const isExpanded = (id: string) => expandedIds.has(id)
-
-    return { isExpanded, toggle }
-}
-
-function FeedbackAndNotesSection({ entries }: { entries: ProposalFeedbackEntry[] }) {
-    const { isExpanded, toggle } = useExpandedEntries(entries)
-
-    return (
-        <Paper p="xl" data-testid="feedback-and-notes-section">
-            <Stack gap="md">
-                <Text fw={600} fz={20}>
-                    Feedback and notes
-                </Text>
-                <Divider />
-                <Stack gap="md" data-testid="feedback-entries">
-                    {entries.map((entry, idx) => (
-                        <Stack key={entry.id} gap="md">
-                            <FeedbackEntry
-                                entry={entry}
-                                isExpanded={isExpanded(entry.id)}
-                                onToggle={() => toggle(entry.id)}
-                            />
-                            {idx < entries.length - 1 && <Divider data-testid="entry-divider" />}
-                        </Stack>
-                    ))}
-                </Stack>
-            </Stack>
-        </Paper>
-    )
 }
 
 function GoToDashboardButton() {
