+++ src/app/[orgSlug]/study/[studyId]/submitted/page.tsx
@@ -1,4 +1,4 @@
-import { getStudyAction } from '@/server/actions/study.actions'
+import { getStudyAction, getProposalFeedbackForStudyAction } from '@/server/actions/study.actions'
 import { getOrgNameFromId } from '@/server/db/queries'
 import { isActionError } from '@/lib/errors'
 import { AlertNotFound } from '@/components/errors'
@@ -13,7 +13,12 @@ export default async function StudySubmittedRoute(props: { params: Promise<{ stu
         return <AlertNotFound title="Study was not found" message="no such study exists" />
     }
 
-    const orgName = await getOrgNameFromId(result.orgId)
+    const [orgName, feedbackResult] = await Promise.all([
+        getOrgNameFromId(result.orgId),
+        getProposalFeedbackForStudyAction({ studyId }),
+    ])
 
-    return <SubmittedView orgSlug={orgSlug} study={result} orgName={orgName} />
+    const entries = isActionError(feedbackResult) ? [] : feedbackResult
+
+    return <SubmittedView orgSlug={orgSlug} study={result} orgName={orgName} entries={entries} />
 }
