+++ src/app/[orgSlug]/study/[studyId]/submitted/submitted-view.tsx
@@ -4,7 +4,7 @@ import { Link } from '@/components/links'
 import { Routes } from '@/lib/routes'
 import { displayOrgName } from '@/lib/string'
 import { PostSubmissionFeatureFlag } from '@/components/openstax-feature-flag'
-import type { SelectedStudy } from '@/server/actions/study.actions'
+import type { ProposalFeedbackEntry, SelectedStudy } from '@/server/actions/study.actions'
 import { StudyRequestPageHeader } from '../../request/page-header'
 import { SubmittedProposalPreview } from './submitted-proposal-preview'
 import { ProposalSubmitted } from './proposal-submitted'
@@ -13,9 +13,10 @@ interface SubmittedViewProps {
     orgSlug: string
     study: SelectedStudy
     orgName: string
+    entries: ProposalFeedbackEntry[]
 }
 
-export function SubmittedView({ orgSlug, study, orgName }: SubmittedViewProps) {
+export function SubmittedView({ orgSlug, study, orgName, entries }: SubmittedViewProps) {
     const existingView = (
         <Stack p="xl" gap="xl">
             <StudyRequestPageHeader orgSlug={orgSlug} studyId={study.id} studyTitle={study.title} />
@@ -43,7 +44,7 @@ export function SubmittedView({ orgSlug, study, orgName }: SubmittedViewProps) {
     return (
         <PostSubmissionFeatureFlag
             defaultContent={existingView}
-            optInContent={<ProposalSubmitted orgSlug={orgSlug} study={study} orgName={orgName} />}
+            optInContent={<ProposalSubmitted orgSlug={orgSlug} study={study} orgName={orgName} entries={entries} />}
         />
     )
 }
